+++ ext/nokogiri/html4_sax_parser_context.c
@@ -77,6 +77,8 @@ noko_html4_sax_parser_context__parse_with(VALUE rb_context, VALUE rb_sax_parser)
    * don't have any cleanup to do. The only memory we need to free is handled by
    * xml_sax_parser_context_type_free */
   htmlParseDocument(ctxt);
+  RB_GC_GUARD(rb_sax_parser);
+  RB_GC_GUARD(rb_context);
 
   return Qnil;
 }