+++ test/test_xslt_transforms.rb
@@ -13,6 +13,54 @@ module Nokogiri
 
       let(:doc) { Nokogiri::XML(File.open(XML_FILE)) }
 
+      [:initialize, :call].each do |failure_point|
+        define_method("test_extension_#{failure_point}_nonlocal_exit_releases_instances") do
+          skip("Ruby extensions are only supported by libxslt") if Nokogiri.jruby?
+          require "weakref"
+
+          instances = []
+          action = nil
+          extension = Class.new do
+            define_method(:initialize) do
+              instances << WeakRef.new(self)
+              action&.call if failure_point == :initialize
+            end
+            define_method(:call) do
+              action&.call if failure_point == :call
+              "success"
+            end
+          end
+          style = Nokogiri::XSLT(<<~XML, "urn:nonlocal-#{failure_point}" => extension)
+            <xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
+                            xmlns:ext="urn:nonlocal-#{failure_point}" extension-element-prefixes="ext">
+              <xsl:strip-space elements="*"/>
+              <xsl:template match="doc"><out><xsl:value-of select="ext:call()"/></out></xsl:template>
+            </xsl:stylesheet>
+          XML
+          document = Nokogiri::XML("<doc> <child/> </doc>")
+          document.root.children.to_a
+          error = RuntimeError.new("expected")
+          action = -> { raise error }
+          assert_same(error, assert_raises(RuntimeError) { style.transform(document) })
+          action = -> { throw(:stop, :done) }
+          assert_equal(:done, catch(:stop) { style.transform(document) })
+          action = lambda do
+            action = -> { raise error }
+            style.transform(document)
+          end
+          assert_same(error, assert_raises(RuntimeError) { style.transform(document) })
+          action = lambda do
+            action = -> { throw(:stop, :nested) }
+            style.transform(document)
+          end
+          assert_equal(:nested, catch(:stop) { style.transform(document) })
+          action = nil
+          assert_equal("success", style.transform(document).root.content)
+          GC.start
+          instances.each { |instance| refute_predicate(instance, :weakref_alive?) }
+        end
+      end
+
       def test_class_methods
         style = Nokogiri::XSLT(File.read(XSLT_FILE))
 
@@ -365,6 +413,35 @@ module Nokogiri
       end
 
       describe "https://github.com/sparklemotion/nokogiri/issues/2800" do
+        it "keeps copied nodes retained by extension callbacks alive" do
+          skip_unless_libxml2("Ruby extensions are only supported by libxslt")
+
+          retained = nil
+          extension = Class.new do
+            define_method(:retain) do |nodes|
+              retained = nodes.first
+              nodes
+            end
+          end
+          style = Nokogiri::XSLT(<<~XML, "urn:retain-copy" => extension)
+            <xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
+                            xmlns:ext="urn:retain-copy" extension-element-prefixes="ext">
+              <xsl:strip-space elements="*"/>
+              <xsl:template match="doc"><out><xsl:value-of select="name(ext:retain(.))"/></out></xsl:template>
+            </xsl:stylesheet>
+          XML
+          document = Nokogiri::XML("<doc> <child/> </doc>")
+          blank = document.root.children.first
+
+          assert_equal("doc", style.transform(document).root.content)
+          refute_same(document, retained.document)
+          GC.start
+          gc_verify_compaction_references if GC.respond_to?(:verify_compaction_references)
+
+          assert_equal("child", retained.children.first.name)
+          assert_equal(" ", blank.content)
+        end
+
         let(:doc) do
           Nokogiri::XML::Document.parse(<<~XML)
             <catalog>